+++ src/test/java/school/redrover/CrazyTestersTest.java
@@ -12,6 +12,8 @@ import school.redrover.runner.BaseTest;
 
 import java.time.Duration;
 
+import static java.sql.DriverManager.getDriver;
+
 public class CrazyTestersTest extends BaseTest {
 
     @Test
@@ -92,4 +94,22 @@ public class CrazyTestersTest extends BaseTest {
         driver.quit();
 
     }
+
+
+    @Test
+    public void testTest() {
+        WebDriver driver = getDriver();
+        driver.manage().timeouts().implicitlyWait(Duration.ofMillis(500));
+        driver.get("https://www.saucedemo.com/");
+
+        WebElement textBox = driver.findElement(By.name("user-name"));
+        textBox.sendKeys("Selenium");
+
+        WebElement login = driver.findElement(By.id("login-button"));
+        login.click();
+
+        WebElement message = driver.findElement(By.cssSelector("h3"));
+        Assert.assertEquals("Epic sadface: Password is required", message.getText());
+    }
 }
+
